[fix] wgan / improved wgan / dcgan architecture.  more
