+++ hypergan/configurations/softmax-gan.json
@@ -1,16 +1,16 @@
 {
   "description": "Softmax GAN",
   "publication": "https://arxiv.org/abs/1704.06191",
+
   "discriminator": 
   {
-    "class": "class:hypergan.discriminators.pyramid_discriminator.PyramidDiscriminator",
-    "layers": 5,
-    "block": "function:hypergan.discriminators.common.standard_block",
-    "initial_depth": 64,
+    "class": "class:hypergan.discriminators.dcgan_discriminator.DCGANDiscriminator",
     "activation": "lrelu",
-    "noise": 1e-2,
-    "depth_increase": 64
+    "final_features": 300,
+    "improved": true,
+    "noise": 1e-2
   },
+
   "encoder":
   {
     "class": "function:hypergan.encoders.uniform_encoder.UniformEncoder",
@@ -22,19 +22,15 @@
     "z": 100
   },
   "generator": {
-    "class": "function:hypergan.generators.resize_conv_generator.ResizeConvGenerator",
-    "activation": "lrelu",
-    "final_activation": "tanh",
-    "depth_increase": 64,
-    "final_depth": 64,
-    "block": "function:hypergan.generators.common.standard_block",
-    "initializer": "random_normal",
-    "random_stddev": 0.02
+    "class": "function:hypergan.generators.dcgan_generator.DCGANGenerator",
+    "activation": "selu"
   },
+
   "loss":
   {
     "class": "function:hypergan.losses.softmax_loss.SoftmaxLoss",
-    "reduce": "reduce_mean"
+    "reduce": "reduce_mean",
+    "gradient_penalty": 10
   },
   "trainer": {
     "class": "class:hypergan.trainers.alternating_trainer.AlternatingTrainer",
